chore: Update UR encoder to version 0.3.2
