+++ interfaces/ur_registry_flutter/lib/ffi/ffi_factory.dart
@@ -1,0 +1,1 @@
+export 'ffi_interface.dart' if (dart.library.io) 'ffi_native.dart';
