+++ interfaces/ur_registry_flutter/lib/ffi/ffi_native.dart
@@ -1,0 +1,2 @@
+export 'dart:ffi';
+export 'package:ffi/ffi.dart';
