+++ interfaces/ur_registry_flutter/lib/native_object.dart
@@ -1,5 +1,4 @@
-import 'dart:ffi';
-
+import 'package:ur_registry_flutter/ffi/ffi_factory.dart';
 import 'package:ur_registry_flutter/ur_registry_flutter.dart';
 
 class NativeObject {
