+++ interfaces/ur_registry_flutter/lib/registries/cardano/cardano_sign_request.dart
@@ -1,12 +1,9 @@
-import 'dart:ffi';
-
-import 'package:ffi/ffi.dart';
+import 'package:ur_registry_flutter/ffi/ffi_factory.dart';
 import 'package:ur_registry_flutter/native_object.dart';
 import 'package:ur_registry_flutter/response.dart';
 import 'package:ur_registry_flutter/ur_encoder.dart';
 import 'package:uuid/uuid.dart';
 import 'package:convert/convert.dart';
-
 const nativePrefix = "cardano_sign_request";
 
 typedef NativeConstruct = Pointer<Response> Function(
