+++ interfaces/ur_registry_flutter/lib/registries/crypto_hd_key.dart
@@ -1,4 +1,4 @@
-import 'dart:ffi';
+import 'package:ur_registry_flutter/ffi/ffi_factory.dart';
 import 'package:ur_registry_flutter/native_object.dart';
 
 import '../response.dart';
