+++ interfaces/ur_registry_flutter/lib/registries/crypto_psbt.dart
@@ -1,7 +1,6 @@
-import 'dart:ffi';
+import 'package:ur_registry_flutter/ffi/ffi_factory.dart';
 
 import 'package:convert/convert.dart';
-import 'package:ffi/ffi.dart';
 import 'package:ur_registry_flutter/native_object.dart';
 import 'package:ur_registry_flutter/response.dart';
