+++ interfaces/ur_registry_flutter/pubspec.yaml
@@ -1,11 +1,11 @@
 name: ur_registry_flutter
 description: An ffi bridge for ur_registry_rust and flutter.
-version: 0.3.1
+version: 0.3.2
 homepage: https://github.com/KeystoneHQ/ur-registry-rust/interfaces/ur_registry_flutter
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=3.3.0 <4.0.0"
+  flutter: ">=3.19.0 <4.0.0"
 
 dependencies:
   convert: ^3.1.1
